+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for further information about branching and tagging conventions.
 
 ## [Unrealised]
 #### Added
+- Telegram chats dropdown
 - Checking is user already exist in submodule
 
 #### Fixed